requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
fonttools==4.47.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.26.3
packaging==23.2
pillow==10.2.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-decouple==3.8
pyttsx3==2.90
pytweening==1.0.7
pywhatkit==5.4
pywin32==306
reportlab==4.0.9
requests==2.31.0
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wikipedia==1.4.0
Wikipedia-API==0.6.0